Possible bug
There is a bug that I couldnt fix.
When getUserPet and getUserTasks are called for the first time, they don't fetch the data fast enough, which causes the page to possibly crash if the page wants to use the date fetched immediately. For tasks this can be solved by fetching the tasks as soon as the main page is loaded, but pet is also shown on the main page. This is a possible problem